requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
python-dotenv
psycopg2-binary
drf-yasg
mysqlclient